Changing prediction datatype

+++ migrations/V004__Prediction_Outputs.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS "weekly_forecast" (
       "id" BIGINT NOT NULL,
       "product_id" BIGINT NOT NULL,
-      "predicted_quantity" BIGINT NOT NULL,
+      "predicted_quantity" NUMERIC(10,2) NOT NULL,
       PRIMARY KEY ("id"),
       CONSTRAINT "fk_product" FOREIGN KEY ("product_id") REFERENCES "products" ("id") ON UPDATE NO ACTION ON DELETE RESTRICT
 );
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS "daily_forecast" (
       "month" INTEGER NOT NULL,
       "day" INTEGER NOT NULL,
       "day_of_week" INTEGER NOT NULL,
-      "predicted_quantity" BIGINT NOT NULL,
+      "predicted_quantity" NUMERIC(10,2) NOT NULL,
       PRIMARY KEY ("id"),
       CONSTRAINT "fk_product" FOREIGN KEY ("product_id") REFERENCES "products" ("id") ON UPDATE NO ACTION ON DELETE RESTRICT
 );
